requirements.txt
fastapi
uvicorn
pydantic-settings
langchain
langchain-community
faster-whisper
python-multipart
httpx
jinja2
aiofiles
sqlalchemy
aiosqlite
# Auth and Security
python-jose[cryptography]
passlib[bcrypt]
bcrypt==4.0.1
# Vector DB
faiss-cpu
numpy
# PDF Processing
pypdf
